@@ -1,68 +1,90 @@
 {
-  "name": "xone_chain_docs",
+  "name": "nextjs-nextra-starter",
+  "type": "module",
   "version": "0.0.1",
-  "description": "Xone Docs",
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "export": "next export"
-  },
+  "description": "Next.js Nextra (v4) Tailwind CSS (v4) docs starter",
+  "author": "Wisdom <pdsu.wwz@foxmail.com>",
+  "license": "MIT",
+  "homepage": "https://github.com/pdsuwwz/nextjs-nextra-starter#readme",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/hello-xone/official_docs.git"
+    "url": "git+https://github.com/pdsuwwz/nextjs-nextra-starter.git"
   },
-  "author": "xone chain <hello@xone.org>",
-  "license": "MIT",
   "bugs": {
-    "url": "https://github.com/hello-xone/official_docs/issues"
+    "url": "https://github.com/pdsuwwz/nextjs-nextra-starter/issues"
   },
-  "homepage": "https://github.com/hello-xone/official_docs#readme",
+  "engines": {
+    "node": ">=20.x",
+    "pnpm": ">=9.x"
+  },
+  "scripts": {
+    "dev": "next dev --turbopack -p 8000",
+    "build": "next build",
+    "postbuild": "next-sitemap --config next-sitemap.config.mjs && pagefind --site .next/server/app --output-path public/_pagefind",
+    "start": "next start -p 7000",
+    "lint": "eslint --fix ."
+  },
   "dependencies": {
-    "@heroui/react": "^2.7.6",
-    "@heroui/toast": "^2.0.7",
-    "@iconify/tailwind": "^1.1.3",
-    "@mantine/code-highlight": "^7.9.1",
-    "@mantine/core": "^7.9.1",
-    "@mantine/hooks": "^7.9.1",
-    "@mantine/notifications": "^7.9.1",
-    "@nextui-org/react": "^2.4.8",
-    "@scalar/api-reference-react": "^0.7.13",
-    "@theguild/remark-mermaid": "^0.2.0",
-    "@vercel/speed-insights": "^1.2.0",
+    "@egoist/tailwindcss-icons": "^1.9.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/server": "^11.11.0",
+    "@emotion/styled": "^11.14.1",
+    "@heroui/react": "^2.8.5",
+    "@mantine/core": "^8.3.7",
+    "@mantine/hooks": "^8.3.7",
+    "@nextui-org/react": "^2.6.11",
+    "@radix-ui/react-accordion": "^1.2.12",
+    "@radix-ui/react-hover-card": "^1.1.15",
+    "@radix-ui/react-separator": "^1.1.7",
+    "@radix-ui/react-slot": "^1.2.3",
+    "@radix-ui/react-toggle": "^1.1.10",
+    "@tsparticles/engine": "^3.9.1",
+    "@tsparticles/react": "^3.0.0",
+    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "date-fns": "^4.1.0",
-    "framer-motion": "^11.18.2",
-    "i18next": "^25.6.0",
-    "libpag": "^4.2.84",
-    "next": "^13.0.6",
-    "next-seo": "^6.6.0",
-    "nextra": "2.13.4",
-    "nextra-theme-blog": "2.13.4",
-    "nextra-theme-docs": "2.13.4",
-    "prism-react-renderer": "^2.4.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-i18next": "^16.2.1",
-    "react-syntax-highlighter": "15.6.1",
-    "sharp": "^0.34.3"
+    "framer-motion": "^12.23.22",
+    "lucide-react": "^0.544.0",
+    "motion": "^12.23.22",
+    "next": "15.5.4",
+    "next-sitemap": "^4.2.3",
+    "next-themes": "^0.4.6",
+    "nextra": "^4.5.1",
+    "nextra-theme-docs": "4.5.1",
+    "qss": "^3.0.0",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-fast-marquee": "^1.6.5",
+    "tailwind-merge": "^3.3.1",
+    "tailwindcss-animate": "^1.0.7",
+    "tsparticles": "^3.9.1"
   },
   "devDependencies": {
-    "@egoist/tailwindcss-icons": "^1.8.1",
-    "@iconify/json": "^2.2.270",
-    "@svgr/webpack": "^8.0.1",
-    "@types/node": "18.11.10",
-    "@types/react": "18.3.12",
-    "@types/react-syntax-highlighter": "15.5.13",
-    "autoprefixer": "^10.4.20",
-    "postcss": "^8.4.49",
-    "tailwindcss": "^3.4.14",
-    "typescript": "^5.8.3"
-  },
-  "pnpm": {
-    "overrides": {
-      "@scalar/code-highlight": "0.1.2"
-    }
-  },
-  "packageManager": "pnpm@10.13.0+sha512.e08cbfa5d0c47e148d9d92965513af3d06a42ff99b6e0e82a7d10512858d0fb680a03c1e6772d369a2305c79306c9257989a3576b7057d6877469c9a1c7563f0"
+    "@antfu/eslint-config": "^5.4.1",
+    "@eslint-react/eslint-plugin": "^2.0.1",
+    "@eslint/compat": "^1.4.0",
+    "@iconify/json": "^2.2.389",
+    "@iconify/tailwind4": "^1.0.6",
+    "@next/third-parties": "15.5.4",
+    "@svgr/core": "^8.1.0",
+    "@svgr/plugin-jsx": "^8.1.0",
+    "@svgr/plugin-svgo": "^8.1.0",
+    "@svgr/webpack": "^8.1.0",
+    "@tailwindcss/postcss": "^4.1.13",
+    "@tailwindcss/typography": "^0.5.19",
+    "@types/node": "^24.5.2",
+    "@types/react": "^19.1.14",
+    "@types/react-dom": "^19.1.9",
+    "eslint": "^9.36.0",
+    "eslint-config-next": "15.5.4",
+    "eslint-plugin-format": "^1.0.2",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.22",
+    "pagefind": "^1.4.0",
+    "postcss": "^8.5.6",
+    "react-responsive": "^10.0.1",
+    "sass": "^1.93.2",
+    "tailwindcss": "^4.1.13",
+    "typescript": "^5.9.2"
+  }
 }
