@@ -1,3 +1,3 @@
-import { createTheme } from "@mantine/core";
+import { createTheme } from '@mantine/core'
 
-export const theme = createTheme({});
+export const theme = createTheme({})
